backend/requirements.txt
fastapi
uvicorn
pypdf
pdfplumber
faiss-cpu
transformers
sentence-transformers
python-multipart
langchain
numpy
torch
scikit-learn